feat: Updated package exports to include types for ComparisonInput and ComparisonResult as well as MatchKey enum
