+++ src/index.ts
@@ -1,1 +1,3 @@
 export { compareWithLabeled } from './compareWithLabeled'
+export { MatchKey } from './constants'
+export type { ComparisonInput, ComparisonResult } from './types'
